Validate that input and output paths can be accessed by snap
Fixes #342 and partially addresses #327 by rejecting attempts to build
or create into or from directories outside of `$HOME` with a clear
error message.

+++ charmtools/generators/generator.py
@@ -57,7 +57,13 @@ class CharmGenerator(object):
         create the files and directories for the new charm.
 
         """
+        home_path = os.path.expanduser('~')
         output_path = self._get_output_path()
+        if home_path == '~':  # expansion failed
+            raise CharmGeneratorException('Could not determine home directory')
+        if not os.path.abspath(output_path).startswith(home_path):
+            raise CharmGeneratorException('Charms can only be created under '
+                                          'your home directory.')
         if os.path.exists(output_path):
             raise CharmGeneratorException(
                 '{} exists. Please move it out of the way.'.format(
@@ -124,4 +130,7 @@ class CharmGenerator(object):
         return tempfile.mkdtemp()
 
     def _cleanup(self, tempdir):
-        shutil.rmtree(tempdir)
+        if os.path.exists(tempdir):
+            # not sure how it could actually get to this point without the
+            # tempdir existing, but we had some reports, so we should check
+            shutil.rmtree(tempdir)
